# openapi.yaml
openapi: 3.0.3
info:
  title: Gestion de Turnos
  description: 'App que permite gestionar turnos para evento unicos y recurrentes para diferentes organizaciones'
  contact:
    email: criscuelieduardo@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /persona:
    post:
      tags:
        - persona
      summary: Agrega una nueva persona
      operationId: addPerson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Persona'
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaPersonaRegistrada'          
        '405':
          description: Entrada invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
    put:
      tags:
        - persona
      summary: Modifica los datos de una persona
      operationId: updatePerson
      parameters:
        - name: dni
          in: header
          description: dni de la persona
          required: true
          schema:
            type: string
        - name: clave
          in: header
          description: clave de la persona
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonaConClave'
        required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'          
        '405':
          description: Entrada invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '412':
          description: Error de identificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
    delete:
      tags:
        - persona
      summary: Baja la cuenta de una persona
      operationId: deletePerson
      parameters:
        - name: dni
          in: header
          description: dni de la persona
          required: true
          schema:
            type: string
        - name: clave
          in: header
          description: clave de la persona
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'          
        '412':
          description: Error de identificacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
    get:
      tags:
        - persona
      summary: Muestra la lista de usuarios registrados
      operationId: getPersons
      parameters:
        - name: apellido
          in: query
          description: filtro por apellidos
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaListaPersonas'          
        '404':
          description: No hay personas registradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
  /persona/{dni}:
    get:
      tags:
        - persona
      summary: Muestra un usuario especifico
      operationId: getPerson
      parameters:
        - name: dni
          in: path
          description: dni de la persona
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaPersona'          
        '404':
          description: No hay personas registradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
  /persona/turno:
    post:
      tags:
        - persona
      summary: Saca un turno para una persona
      operationId: addTurn
      parameters:
        - name: cuit
          in: header
          description: cuit de la organizacion
          required: true
          schema:
            type: string
        - name: evento
          in: header
          description: nombre del evento
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Turno'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '404':
          description: Evento inexistente para la organizacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '405':
          description: Entrada invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Respuesta'
components:
  schemas:
...
    Persona:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: number
    PersonaConClave:
      type: object
      allOf:
        - $ref: '#/components/schemas/Persona'
      properties:
        clave: 
          type: string
    Turno:
      type: object
      properties:
        fecha:
          type: string
        hora:
          type: string
        persona:
          $ref: '#/components/schemas/Persona'
    Respuesta:
      type: object
      properties:
        status:
          type: number
        descripcion:
          type: string
    RespuestaPersonaRegistrada:
      type: object
      allOf:
        - $ref: '#/components/schemas/Respuesta'
      properties:
        clave:
          type: string
    RespuestaListaPersonas:
      type: object
      allOf:
        - $ref: '#/components/schemas/Respuesta'
      properties:
        personas:
          type: array
          items:  
            $ref: '#/components/schemas/PersonaConClave'
    RespuestaPersona:
      type: object
      allOf:
        - $ref: '#/components/schemas/Respuesta'
      properties:
        persona:
          $ref: '#/components/schemas/PersonaConClave'
...